@@ -2,14 +2,19 @@ apply plugin: 'com.android.application'
 apply plugin: 'kotlin-android'
 
 android {
-	compileSdkVersion 27
+	compileSdkVersion 28
 	defaultConfig {
 		applicationId "com.sensorberg.synchronousgatt.sample"
 		minSdkVersion 23
-		targetSdkVersion 27
+		targetSdkVersion 28
 		versionCode 1
 		versionName "1.0"
-		testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+		testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
+
+		buildConfigField "String", "SERVICE_UUID", findQuotedProperty("SERVICE_UUID", null)
+		buildConfigField "String", "CHARACTERISTIC_WRITE_UUID", findQuotedProperty("CHARACTERISTIC_WRITE_UUID", null)
+		buildConfigField "String", "CHARACTERISTIC_READ_UUID", findQuotedProperty("CHARACTERISTIC_READ_UUID", null)
+		buildConfigField "String", "DESCRIPTOR_NOTIFY_UUID", findQuotedProperty("DESCRIPTOR_NOTIFY_UUID", wrapInQuotes("00002902-0000-1000-8000-00805f9b34fb"))
 	}
 	buildTypes {
 		release {
@@ -21,11 +26,11 @@ android {
 
 dependencies {
 	implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
-	implementation project(':library')
+	implementation project(':synchronousgatt')
 	implementation "com.jakewharton.timber:timber:4.7.1"
-	implementation 'com.android.support:appcompat-v7:27.1.1'
-	implementation 'com.android.support:recyclerview-v7:27.1.1'
-	implementation 'com.sensorberg.libs:permission-bitte:0.2.2'
-	implementation 'no.nordicsemi.android.support.v18:scanner:1.1.0'
-	implementation 'com.android.support.constraint:constraint-layout:1.1.3'
+	implementation 'androidx.appcompat:appcompat:1.1.0'
+	implementation 'androidx.recyclerview:recyclerview:1.0.0'
+	implementation 'com.sensorberg.libs:permission-bitte:0.5.0'
+	implementation 'no.nordicsemi.android.support.v18:scanner:1.4.3'
+	implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
 }
